requirements.txt
Flask
langchain-ibm
ibm-watsonx-ai
langchain
langchain-community
chromadb
python-dotenv
gunicorn
tiktoken
pydantic 